moving stuff from `TracedTurns` to `TelemetryReader`
