+++ akka-javasdk-testkit/src/main/java/akka/javasdk/testkit/eval/EvalTarget.java
@@ -4,6 +4,7 @@
 
 package akka.javasdk.testkit.eval;
 
+import akka.javasdk.testkit.ToolCall;
 import java.util.List;
 
 /**
