+++ akka-javasdk-testkit/src/main/java/akka/javasdk/testkit/eval/Judge.java
@@ -4,6 +4,7 @@
 
 package akka.javasdk.testkit.eval;
 
+import akka.javasdk.testkit.ToolCall;
 import akka.javasdk.testkit.eval.Evaluator.EvalResult;
 import java.util.List;
 import java.util.Locale;
