+++ akka-javasdk-testkit/src/main/java/akka/javasdk/testkit/eval/package-info.java
@@ -12,7 +12,7 @@
  * <p>{@link akka.javasdk.testkit.eval.ExperimentRunner} calls the agent in a fresh session per
  * case. Run one case at a time with a mocked model, or a whole batch judged by a {@link
  * akka.javasdk.testkit.eval.Gate} with a real model. Tool and model evidence is read from the
- * runtime trace through {@link akka.javasdk.testkit.eval.TracedTurns}.
+ * runtime trace through {@link akka.javasdk.testkit.TelemetryReader}.
  *
  * <p>Criteria the built-in expectations cannot express, such as tone or completeness, go to a model
  * through {@link akka.javasdk.testkit.eval.Judge}.
